hook: revamp the name-status to work with amending
The 'prepare-commit-msg' hook has a sample that added
the name-status output of diff to the commit message.
It's commented out as it didn't cope with amending
and squashing. Further, the hook deosn't have any other
useful examples to show what could be done with it.

Revamp the sample to make it work with amending alone.
It needed a little script hackery along with the help
of tools such as 'sort'.

With respeect to amending, it currently works only with
basic combinations of changes to same files in the last
commit and in the current index as given by the following
table,

      Combinations of change of status to same file
           while amending and their results
                 (not exhaustive)

    -------------------------------------------------
    |  Last commit   |   Index     |    Result      |
    -------------------------------------------------
    |      A         |     M       |       ✓        |
    |      A         |     D       |       ✕        |
    |      A         |     R       |       ✕        |
    |      D         |     A       |       ✕        |
    |      D         |     M       |       ✓        |
    |      D         |     R       |       ✕        |
    |      M         |     A       |       NA       |
    |      M         |     D       |       ✕        |
    |      M         |     R       |       ✕        |
    |      R         |     A       |       ✕        |
    |      R         |     M       |       ✕        |
    |      R         |     D       |       NA       |
    ------------------------------------------------

There are other status combinations (e.g. those includein copying)
that have not been tested for. And of course, these results are only
for status combinations of same files. It's worths depends on how
much you consider the name status of changes to same files while
amending.

Signed-off-by: Kaartic Sivaraam <kaarticsivaraam91196@gmail.com>