commit-template: change a message to be more intuitive
It's not good to use the phrase 'do not touch' to convey the information
that the cut-line should not be modified or removed as it could possibly
be mis-interpreted by a person who doesn't know that the word 'touch' has
the meaning of 'tamper with'. Further, it could make translations a little
difficult as it might not have the intended meaning in a few languages (for
which translations don't exist yet) when translated as such.

So, use intuitive terms in the sentence. Replacing the word 'touch' with
other terms has introduced the possibility of the following sentence to be
mis-interpreted, so change the terms in that too.

Signed-off-by: Kaartic Sivaraam <kaarticsivaraam91196@gmail.com>