branch: introduce parameter to prevent branchname validation from die()ing
This parameter allows the branchname validation functions to
optionally return a flag specifying the reason for failure, when
requeste. This allows the caller to know why it was about to die as
it now "gently" returns a value indicating the error without die()ing
directly. This allows more useful error messages to be given to the
user when trying to rename a branch.

The flags are specified in the form of an enum and values for success
flags have been assigned explicitly to clearly express that certain
callers rely on those values and they cannot be arbitrary.

Only the logic has been added but no caller has been made to use
it, yet. So, no functional changes.

Signed-off-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>