build: add alias for global build correctly
A variable required to add an alias for the global build was removed
carelessly in one of the commits which broke the setup.

Specify the build location in the script itself for the sake of
consistency. This time the instal location uses "${HOME}" instead
of $HOME which gives the user the flexibility to move the HOME
directory without breaking the alias.

Signed-off-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>