rebase: use a more appropriate variable name
The variable name "branch_name" used to indicate the <branch> parameter
in "git rebase <upstream> <branch>" is misleading as it seems to indicate
that it holds only a "branch name" although at times it might actually hold
any valid <ref> (branch/commit).

So, use a more suitable name for that variable.

Signed-off-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>