Doc/git-remote: avoid confusion w.r.t *-url behaviour
The documentation were very clear in specifying indirectly that
"git remote *-url" by default manipulate only fetch URLs
(remote.$remote_name.url). Unfortunately this could lead to confusion
...

MAY BE NOT: This doesn't sound as confusing as it seems. If we do this
we would possibly be losing the only info to say that the fetch and
push URLs might be different! So, let's drop this.

Signed-off-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>