Try eliminating find_unique_abbrev()
Try eliminating find_uniqe_abbrev() by passing the commit hash directly
to print_submodule_summary() instead of calculating the hash using the
former.

Quote from Dscho:
That is not quite the same, as it looks for the abbreviation _in the
submodule_, not in the current project. So I think `find_unique_abbrev()`
is not correct here.

The funny thing is that we _already_ will have called `git rev-parse
--verify` for both `p->oid_src` and `p->oid_dst` in the submodule, in the
caller of this function! And while we throw away the result, and while we
do not pass `--short`, there is no reason why we shouldn't be able to do
precisely that.

Quoting Dscho:

    It looks as if there is a whole lot of `oid_to_hex(&p->oid_src)` in that
    function. Together with the realization that we need the abbreviated
    version of that at least in one place, and the other realization that we
    already call `rev-parse --verify` for both `oid_src` and `oid_dst` in the
    caller of this function, it seems to suggest itself that we would actually
    want to pass the `--short` option, too, and to capture the output, and
    pass it down to `print_submodule_summary()` _instead of_ `missing_src` and
    `missing_dst` (e.g. as `src_abbrev` and `dst_abbrev`).
--------------------------------------------------------------------

The whole mechanism of finding 'total_commits' relies on src_oid_hash
and dst_oid_hash being non-NULL. If verify_submodule_committish()
returns a NULL. This means that the aforementioned char(s) have to be made
"0000000". I am trying to do this yet I am getting very absurd errors
such as:
    free(): invalid pointer
    ----------
    sm1 does not contain the commit <seemingly valid SHA>

I do not know what is going on. I googled the first error, seems that
you cannot free stack memory but rather only heap memory (the one
allocated via malloc() et al.).

About the second one, I don't know how did it come up. A solution I
tried was to make the variables "0000000" if they are still NULL but I
had no luck. Also, for seemingly valid SHA, the error that came up was:

    Warn: <submodule-name> doesn't contain commit <SHA>

Any suggestions on this?

THIS IS BUILT ON summary-v2_5 therefore has the latest changes.

https://lore.kernel.org/git/nycvar.QRO.7.76.6.2007031712160.50@tvgsbejvaqbjf.bet/